+++ CHANGES.rst
@@ -28,6 +28,7 @@ master (in development)
   - Improve GCC syntax checking by expanding templates [GH-459]
   - `d-dmd` reports errors with columns now [GH-460]
   - Remove Projectile-based config file search [GH-461]
+  - Do not change point when navigating in the error list [GH-487]
 
 - Bug fixes: